elasticsearch 部分程序

+++ src/main/java/com/example/hemahotel/entity/Hotel.java
@@ -32,7 +32,7 @@ public class Hotel {
     //private Double price;//酒店房间中最低价
 
     @Column
-    private double star;//星级
+    private Integer star;//星级
 
     @Column
     private String phone;//酒店客服联系电话
